+++ build.sbt
@@ -8,7 +8,7 @@ lazy val zioVersion           = "2.0.2"
 lazy val kafkaVersion         = "3.2.0"
 lazy val embeddedKafkaVersion = "3.2.0" // Should be the same as kafkaVersion, except for the patch part
 
-lazy val embeddedKafka = "io.github.embeddedkafka" %% "embedded-kafka" % embeddedKafkaVersion % Test
+lazy val embeddedKafka = "io.github.embeddedkafka" %% "embedded-kafka" % embeddedKafkaVersion
 
 inThisBuild(
   List(
@@ -82,20 +82,10 @@ lazy val zioKafka =
     .settings(
       libraryDependencies ++= Seq(
         "dev.zio"                   %% "zio-streams"             % zioVersion,
-        "dev.zio"                   %% "zio-test"                % zioVersion % Test,
-        "dev.zio"                   %% "zio-test-sbt"            % zioVersion % Test,
         "org.apache.kafka"           % "kafka-clients"           % kafkaVersion,
         "com.fasterxml.jackson.core" % "jackson-databind"        % "2.13.3",
-        "ch.qos.logback"             % "logback-classic"         % "1.2.11"   % Test,
         "org.scala-lang.modules"    %% "scala-collection-compat" % "2.7.0"
-      ) ++ {
-        if (scalaBinaryVersion.value == "3")
-          Seq(
-            embeddedKafka
-              .cross(CrossVersion.for3Use2_13) exclude ("org.scala-lang.modules", "scala-collection-compat_2.13")
-          )
-        else Seq(embeddedKafka)
-      }
+      )
     )
 
 lazy val zioKafkaTestUtils =
