--- zio-kafka-test/src/test/scala/zio/kafka/embedded/Kafka.scala
@@ -1,31 +1,0 @@
-package zio.kafka.embedded
-
-import io.github.embeddedkafka.{ EmbeddedK, EmbeddedKafka, EmbeddedKafkaConfig }
-import zio._
-
-trait Kafka {
-  def bootstrapServers: List[String]
-  def stop(): UIO[Unit]
-}
-
-object Kafka {
-
-  final case class EmbeddedKafkaService(embeddedK: EmbeddedK) extends Kafka {
-    override def bootstrapServers: List[String] = List(s"localhost:${embeddedK.config.kafkaPort}")
-    override def stop(): UIO[Unit]              = ZIO.succeed(embeddedK.stop(true))
-  }
-
-  case object DefaultLocal extends Kafka {
-    override def bootstrapServers: List[String] = List(s"localhost:9092")
-    override def stop(): UIO[Unit]              = ZIO.unit
-  }
-
-  val embedded: ZLayer[Any, Throwable, Kafka] = ZLayer.scoped {
-    implicit val embeddedKafkaConfig: EmbeddedKafkaConfig = EmbeddedKafkaConfig(
-      customBrokerProperties = Map("group.min.session.timeout.ms" -> "500", "group.initial.rebalance.delay.ms" -> "0")
-    )
-    ZIO.acquireRelease(ZIO.attempt(EmbeddedKafkaService(EmbeddedKafka.start())))(_.stop())
-  }
-
-  val local: ZLayer[Any, Nothing, Kafka] = ZLayer.succeed(DefaultLocal)
-}
